Services/auth_service/requirements.txt
Django==5.0.6
djangorestframework==3.14.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
django-rest-passwordreset==1.4.0